changed delay to double

+++ Memoizer/CacheAttribute.cs
@@ -10,12 +10,12 @@ namespace Memoizer
     [Injection(typeof(MemoizeAspect))]
     public class CacheAttribute : Attribute
     {
-        private readonly int? MilliSeconds;
+        private readonly double? MilliSeconds;
         public CacheAttribute()
         {
             MilliSeconds = null;
         }
-        public CacheAttribute(int expiry, Time unit = Time.Millisecond)
+        public CacheAttribute(double expiry, Time unit = Time.Millisecond)
         {
             if (unit == Time.Second)
             {
@@ -27,6 +27,9 @@ namespace Memoizer
             }
         }
 
-        public DateTimeOffset? ExpiryFromNow => MilliSeconds == null ? null : (DateTimeOffset?)DateTimeOffset.UtcNow.AddMilliseconds((int)MilliSeconds);
+        public DateTimeOffset? ExpiryFromNow =>
+            MilliSeconds == null ? 
+            null : 
+            (DateTimeOffset?)DateTimeOffset.UtcNow.AddMilliseconds((double)MilliSeconds);
     }
 }
